Jpa funzionante

+++ OnlineValidator/src/main/java/com/onlinevalidator/repository/TipoDocumentoJpaRepositoryInterface.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.onlinevalidator.model.Tipodocumento;
 
-public interface TipoDocumentoJpaRepositoryInterface extends JpaRepository<Tipodocumento, Long>{
+public interface TipoDocumentoJpaRepositoryInterface extends JpaRepository<Tipodocumento, Integer>{
 
 }
